+++ libSBML API/Species/Species_setInitialAmount.m
@@ -73,9 +73,9 @@ elseif(~isnumeric(initialAmount))
 end;
 
 SBMLSpecies.initialAmount = initialAmount;
-SBMLSpecies.isSetInitialAmount = 1;
+SBMLSpecies.isSetInitialAmount = int32(1);
 
 warning off all;
 SBMLSpecies.initialConcentration = 0/0;
-SBMLSpecies.isSetInitialConcentration = 0;
+SBMLSpecies.isSetInitialConcentration = int32(0);
 warning on all;
